+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/settings/InMemorySettings.kt
@@ -69,6 +69,7 @@ class InMemorySettings(project: Project) {
         bootstrapSize = DEFAULT_SIZE
         remixSize = DEFAULT_SIZE
         boxiconsSize = DEFAULT_SIZE
+        fontAwesomeSize = DEFAULT_SIZE
     }
 
     fun updateUIState(uiState: SavedState) {